Updates.
Added a few words to /reset message.

Disabled HUD.cs, it isn't of any use at the moment as it is for an old
HUD I was working on.

Fixed(?) the fact that when you place a brick, it takes from the brick
value, but when you hammer it, it doesn't give it back.